+++ lib/faastruby/cli/commands/function/update_context.rb
@@ -35,8 +35,12 @@ module FaaStRuby
         def usage
           puts "\nUsage: faastruby #{self.class.help}"
           puts %(
--d, --data 'STRING'    # The context data. Must be quoted.
---stdin                # Read context data from STDIN
+-d, --data JSON_STRING'     # The data to be stored as execution context
+                            #  in the cloud, accessible via 'event.context'
+                            #  from within your function.
+                            #  The context data must be a JSON String, and
+                            #  have maximum size of 4KB.
+--stdin                     # Read context data from STDIN
           )
         end