+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Remove unused arguments from update workspace `run` method.
 - Update Help
 - Change the argument that measures the execution time on `run` to `--measure`
+- Add argument to read context from STDIN when running `deploy-to`
 
 ## 0.5.15 - Mar 18 2019
 - Remove `nil` entries from error array before trying to print error messages