+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add argument to read context from STDIN when running `deploy-to`
 - Switch default config files method from array to multiline string
 - Allow passwords with up to 50 characters.
+- Fix bug preventing `faastruby deploy` to deploy the secrets.
 
 ## 0.5.15 - Mar 18 2019
 - Remove `nil` entries from error array before trying to print error messages